pyproject.toml
[project]
name = "plagiarism-remover"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.121.0",
    "groq>=0.33.0",
    "langchain>=1.0.4",
    "langchain-core>=1.0.3",
    "langchain-groq>=1.0.0",
    "langgraph>=1.0.2",
    "pymupdf>=1.26.6",
    "python-dotenv>=1.2.1",
    "simhash>=2.1.2",
    "streamlit>=1.51.0",
]
